pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tapps-agents"
version = "2.0.0"
description = "A specification framework for defining, configuring, and orchestrating coding agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Tapps"}
]
maintainers = [
    {name = "Tapps"}
]
keywords = ["ai", "agents", "coding", "workflow", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.12.5",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "aiohttp>=3.13.2",
    "psutil>=5.9.0",
    "radon>=6.0.1",
    "bandit>=1.9.2",
    "pylint>=4.0.4",
    "coverage>=7.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
    "black>=25.12.0",
    "ruff>=0.14.8,<1.0",
    "mypy>=1.19.0,<2.0",
    "pip-audit>=2.10.0",
    "pipdeptree>=2.30.0",
    "jinja2>=3.1.6",
    "plotly>=6.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.12.0",
    "ruff>=0.14.8,<1.0",
    "mypy>=1.19.0,<2.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
]

[project.urls]
Homepage = "https://github.com/wtthornton/TappsCodingAgents"
Repository = "https://github.com/wtthornton/TappsCodingAgents"
Documentation = "https://github.com/wtthornton/TappsCodingAgents"
Issues = "https://github.com/wtthornton/TappsCodingAgents/issues"

[project.scripts]
tapps-agents = "tapps_agents.cli:main"

[tool.setuptools]
packages = ["tapps_agents"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.md", "*.j2"]

[tool.black]
line-length = 88
target-version = ["py310"]
exclude = '''
/(
  \.git
| \.mypy_cache
| \.pytest_cache
| \.ruff_cache
| \.venv
| build
| dist
| htmlcov
| MagicMock
| tapps_agents\.egg-info
| reports
)/'''

[tool.ruff]
line-length = 88
target-version = "py310"
fix = false
show-fixes = true
exclude = [
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  ".ruff_cache",
  ".venv",
  "build",
  "dist",
  "htmlcov",
  "MagicMock",
  "tapps_agents.egg-info",
  "reports",
]

[tool.ruff.lint]
# Start with a pragmatic, high-signal rule set; tighten over time during the marathon.
select = ["E", "F", "I", "UP", "B"]
ignore = [
  # Let Black handle line-length; long strings/URLs are often intentional.
  "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  # Tests often have “unused” fixtures/parameters by design.
  "ARG001",
  "ARG002",
]

[tool.mypy]
python_version = "3.10"
pretty = true
show_error_codes = true
show_column_numbers = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = false
warn_unreachable = true
no_implicit_optional = true
disallow_untyped_defs = false
check_untyped_defs = false
ignore_missing_imports = true
exclude = "(^build/|^dist/|^htmlcov/|^MagicMock/|^reports/|^tapps_agents\\.egg-info/)"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
  "-q",
  "--disable-warnings",
  "--maxfail=1",
]
asyncio_mode = "auto"
timeout = 120

[tool.coverage.run]
branch = true
source = ["tapps_agents"]

[tool.coverage.report]
skip_empty = true
show_missing = true
precision = 2